refactor(LottoRank) : LottoRank 패키지 이동
issue #427
