+++ src/main/java/lotto/dto/LottoStatisticsDto.java
@@ -1,5 +1,7 @@
 package lotto.dto;
 
+import lotto.domain.LottoRank;
+
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashMap;
