+++ src/test/java/lotto/domain/LottoCheckerTest.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import lotto.dto.LottoNumberDto;
-import lotto.dto.LottoRank;
 import lotto.dto.LottoStatisticsDto;
 import org.junit.jupiter.api.Test;
