+++ src/test/java/lotto/domain/LottoWinningNumberTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.dto.LottoRank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
