+++ src/test/java/lotto/dto/LottoRankTest.java
@@ -1,5 +1,6 @@
 package lotto.dto;
 
+import lotto.domain.LottoRank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
